+++ app/src/components/ShowTsuraiTimeline.vue
@@ -46,9 +46,9 @@
 
 <script>
 import Icon from 'vue-awesome/components/Icon'
-import firebase from 'firebase'
 import Stamp from '@/components/Stamp'
-const firestore = firebase.firestore()
+import { postDeletable } from '@/mixins/postDeletable'
+import { stampHandler } from '@/mixins/stampHandler'
 
 export default {
   name: 'Timeline',
@@ -56,6 +56,10 @@ export default {
     'v-icon': Icon,
     'stamp': Stamp
   },
+  mixins: [
+    postDeletable,
+    stampHandler
+  ],
   data () {
     return {}
   },
@@ -93,83 +97,6 @@ export default {
     isAuthor () {
       if (this.user_id === undefined) return false
       return this.user_id === this.post.author.user_id
-    }
-  },
-  methods: {
-    toggle_too_bad () {
-      if (!this.$store.state.isLoggedIn) {
-        alert('スタンプを押すにはログインしてください')
-        return
-      }
-      const newRef = firestore.doc('/users/' + this.user_id)
-      const idList = this.post.posts.too_bad_id_list
-      const index = idList.findIndex(item => item === this.user_id)
-      if (index === -1) {
-        firestore.collection('posts').doc(this.post.posts.id).update({
-          that_is_too_bad_list: firebase.firestore.FieldValue.arrayUnion(newRef)
-        })
-      } else {
-        firestore.collection('posts').doc(this.post.posts.id).update({
-          that_is_too_bad_list: firebase.firestore.FieldValue.arrayRemove(newRef)
-        })
-      }
-    },
-    toggle_alright () {
-      if (!this.$store.state.isLoggedIn) {
-        alert('スタンプを押すにはログインしてください')
-        return
-      }
-      const newRef = firestore.doc('/users/' + this.user_id)
-      const idList = this.post.posts.alright_id_list
-      const index = idList.findIndex(item => item === this.user_id)
-      if (index === -1) {
-        firestore.collection('posts').doc(this.post.posts.id).update({
-          you_are_alright_list: firebase.firestore.FieldValue.arrayUnion(newRef)
-        })
-      } else {
-        firestore.collection('posts').doc(this.post.posts.id).update({
-          you_are_alright_list: firebase.firestore.FieldValue.arrayRemove(newRef)
-        })
-      }
-    },
-    toggle_good_job () {
-      if (!this.$store.state.isLoggedIn) {
-        alert('スタンプを押すにはログインしてください')
-        return
-      }
-      const newRef = firestore.doc('/users/' + this.user_id)
-      const idList = this.post.posts.good_job_id_list
-      const index = idList.findIndex(item => item === this.user_id)
-      if (index === -1) {
-        firestore.collection('posts').doc(this.post.posts.id).update({
-          good_job_list: firebase.firestore.FieldValue.arrayUnion(newRef)
-        })
-      } else {
-        firestore.collection('posts').doc(this.post.posts.id).update({
-          good_job_list: firebase.firestore.FieldValue.arrayRemove(newRef)
-        })
-      }
-    },
-    isActive (stampType) {
-      if (this.user_id === undefined) return false
-      switch (stampType) {
-        case 'too-bad':
-          return this.post.posts.too_bad_id_list.includes(this.user_id)
-        case 'alright':
-          return this.post.posts.alright_id_list.includes(this.user_id)
-        case 'good-job':
-          return this.post.posts.good_job_id_list.includes(this.user_id)
-      }
-      return false
-    },
-    deletePost () {
-      console.log(this.post.posts_id)
-      if (confirm('投稿を削除しますか?')) {
-        firestore.collection('posts').doc(this.post.posts.id).delete().then(() => {
-          alert('投稿が削除されました')
-        })
-          .catch(e => alert('投稿を削除できませんでした。管理者にお問い合わせください。'))
-      }
     }
   }
 }
